fixes #4
In Android M & N, the `ArtMethod` was hiden in native framework, we can not direct acces it from Java world.

The field `artMethod`  of `java.lang.reflect.AbstractMethod`,is just a long type in Android M & N, if we just copy the origin artMethod, we can not got a new ArtMethod in native--we just copy a long value.

Under Android M, when we call the constructor of `java.lang.reflect.ArtMethod`, we can get a memory block of ArtMethod, it can be access from native world. But above Androd M, the ArtMethod can not allocate from Java world.

so, we use DirectByteBuffer alloc a native heap memoy, which is important, the ArtMethod belongs to the NonMovable Object of JVM, it must not be affected by gc. then we copy the origin method's artMethod memory to the new allocated bytebuffer.

And another thing, we must modify the access flag of the method to `private` to make the new allocate method a `direct method`, Under Android M, we just change the accessFlag of ArtMethod through Reflection, but In Android M, the access flags used by native framework is the native mirror::ArtMethod, we can not just modify the AbstractMethod's accessFlags field. so, we modify the new alloct memory block.