[Fix] fixes #8
When the origin method throw an exception, we can not eat it simply. Because some gays use exception to control the logic of code. If we just catch the exception and do nothing, we may influence the control flow delicately. In a word, when origin method throws exception, we just rethrow it to ensure the right flow.