[Opt]: Explicit is better than implicit.
if we compare a long type with a int type, (0xFFFFFFFF is an integer, while 0xFFFFFFFFL is a long) the long value is cast to int, and then perform the comparasion.
So, when we use arithmetic shift, the real highest bytes is incorrect( negative is shift with sign bit, -0x1122L >> 32 = 0xffffffffffffffffL), but we compare it with an integer, so the result is correct.

Instead, we should use explicit logical shift and compare with a long type.